Interpret pod security context IDs as container IDs instead of host IDs.
Small customization in CRI-O to improve container user-ns isolation; we plan to
work with the CRI-O team to upstream this (or a similar) change.

Problem: for user-ns mappings, CRI-O currently interprets the UID & GIDs in the
pod's spec security context as host IDs rather than container IDs. For example,
"runAsUser: 1000" causes CRI-O to map the root user in the container to host ID
1000, and "runAsUser: 0" causes CRI-O to map the root user in the container to
host ID 0. This is not ideal for container isolation for obvious reasons, and it
also makes it harder on users since they now need to worry about host ID
mappings. But it was done by design on CRI-O given that filesystem user ID
shifting/remapping is not yet widely available.

Solution: for low-level container runtimes such as Sysbox that support
filesystem user ID shifting/remapping, we don't need the pod's security context
to refer to Host IDs. In fact, we want the pod's security context to refer to
container IDs instead, as this is easier for users to understand and provides
stronger container isolation (i.e., users only worry about IDs in the container
and let the underlying runtime choose the proper user-ns host ID mappings to
ensure container isolation).

For example, "runAsUser: 1000" should cause CRI-O to start all containers in the
pod as user 1000 within the container, and map all user IDs in the container to
a range of unprivileged host IDs chosen by CRI-O. Similarly, "runAsUser: 0"
should cause CRI-O to start all containers in the pod as user 0 within the
container and map all user IDs in the container to unprivileged host IDs chosen
by CRI-O.

This commit applies the solution described above.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>